BOO-981: Updated how keyfiles are specified to the compilers, fixed some obsolete issues regarding Hash.cs and BooHashCodeProvider.cs and equality comparers, commented out blocks of code that were intentionally unreachable.
Unfortunately, there are a bunch of antlr warnings, so there are still warnings in a full build, but there are now almost none when warm-building after ast-generation (excepting Boo.Lang.Useful).

Also, Boo.Lang's handling of keyfiles cannot be updated due to a bug in NAnt's (v0.85) csc task implementation.  The bug causes the main assembly to be signed, but doesn't sign the satellite assemblies.  This is fixed in NAnt 0.86-beta1.