WORK IN PROGRESS - Massive syntax and AST change to support type references like List[of int].Enumerator.
SimpleTypeReferences are now chained to each other, each representing a single link in a qualified type name and can be bound to either a type or a namespace. NameResolutionService and other parts changed to reflect this.
Most bugs ironed out, some tests still failing. 

PLEASE, take a look at this. If you disagree with the direction I'm taking here, please say so.

Issue(s): 853